Restructure configuration controller & test files, add better message for configuration state
