+++ api/v1beta2/configuration_types.go
@@ -17,7 +17,6 @@ limitations under the License.
 package v1beta2
 
 import (
-	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 
@@ -78,7 +77,7 @@ type ConfigurationSpec struct {
 	ForceDelete *bool `json:"forceDelete,omitempty"`
 
 	// GitCredentialsReference specifies the reference to the secret containing the git credentials
-	GitCredentialsReference *v1.SecretReference `json:"gitCredentialsReference,omitempty"`
+	GitCredentialsReference *types.SecretReference `json:"gitCredentialsReference,omitempty"`
 }
 
 // ConfigurationStatus defines the observed state of Configuration
