+++ api/v1beta2/zz_generated.deepcopy.go
@@ -23,7 +23,6 @@ package v1beta2
 
 import (
 	crossplane_runtime "github.com/oam-dev/terraform-controller/api/types/crossplane-runtime"
-	"k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 )
 
@@ -188,7 +187,7 @@ func (in *ConfigurationSpec) DeepCopyInto(out *ConfigurationSpec) {
 	}
 	if in.GitCredentialsReference != nil {
 		in, out := &in.GitCredentialsReference, &out.GitCredentialsReference
-		*out = new(v1.SecretReference)
+		*out = new(crossplane_runtime.SecretReference)
 		**out = **in
 	}
 }
