+++ chart/crds/terraform.core.oam.dev_configurations.yaml
@@ -280,13 +280,13 @@ spec:
                   secret containing the git credentials
                 properties:
                   name:
-                    description: Name is unique within a namespace to reference a
-                      secret resource.
+                    description: Name of the secret.
                     type: string
                   namespace:
-                    description: Namespace defines the space within which the secret
-                      name must be unique.
+                    description: Namespace of the secret.
                     type: string
+                required:
+                - name
                 type: object
               hcl:
                 description: HCL is the Terraform HCL type configuration
